Added ctl.sh compile command and running it during ctl.sh deploy (#28)
* Added ctl.sh compile command and running it during ctl.sh deploy

@secult has raised a problem about permissions of byte-compiled
modules (.pyc) in ERIGONES_HOME in production. The source and compiled
files are owned by root, but are accessed by processes that run under
the erigones user. After an upgrade of source files the pyc files
are not updated, because the services do not have permissions
to re-compile the bytecode.