Updated set_request_method() and call_api_view() functions
These functions are used whenever the code needs to call another view.
Usually, the `call_api_view()` function is used when performing internal API calls or by every call made via socket.io. The `set_request_method()` is sometimes needed when the when the caller does not want to call the whole view function, but only a smaller part of the view.

In both cases, the call to the new view should use a copy of the request object. The request object in the current function should not be modified by `set_request_method()` or `call_api_view()`. I've changed the `call_api_view()` so it always uses a copy of the request object. The `set_request_method()` also creates a copy of the request object (except when called from `call_api_view()`) and therefore there is no need to set the old method back.

The race condition was observed during `DELETE node_image_list`. The `NodeImageView.cleanup()` method calls `call_api_view()` multiple times in a for loop and every function call uses one request object. When the API view is trigger from GUI, the very same request object can be altered in an another call made from socket.io. I believe that this not the only use case and many more race conditions can be avoided with this simple fix.