Added None value for OS Type if it does not match supported list.
Due to error in import, when xls file contained OS Type that is not
defined in our system user got 500 error. I have added value to None if
it doesnt match anything. I have also added break statement so when OS
Type match is found we dont need to look throught the rest of the list
any more.