Moved Create DNS checkbox to non advanced section when creating (editing) NIC in VM issue #145
Reason for doing this is that we do create DNS record automatically for the first NIC that is added to VM. Which is a bit magic for user if he doesn't see this check box. If he removes the NIC and than try to re-create no DNS record is created.