Fixed VM hostname fetching in `message_callback` (GUI/JS)
After introduction of hostname_or_uuid for VMs (#16) the
`message_callback` had received the hostname attribute in a
`hostname_or_uuid` field. In case of a started task the hostname field
will be stored in the apiview - so the apiview.hostname should be used a
primary source of VM hostname.