Updated vms model with manifest_path and api to use it issue #155
Updated vms model to store path to VM manifest stored with the backup,
added help functions to generate the paths to be consistent with
backups. **Added migration for this change**.

Added settings for default paths to manifests.

Updated api to be able to pass data from vms model to esbackup script,
so we can append to functionality done in esbackup script.