Do not display *pending* status when the target VM status was already reached (#152)
* Moved `vm.update_uptime()` closer to new VM status saving

It did not make sense to update the uptime and not saving it. This
happened when a new VM state with an old change time propagated to
`_save_vm_status()` function.

* Updated `vm.status_display()` to not display *pending* status when the target VM status was already reached

There is a following race condition visible in the GUI:

- start a stopped VM
- after the VM status changes to running switch to the console or snapshot view
- the VM may be now in pending state (blue icon in the VM menu)

This can happen because a new VM status is delivered from the compute
node to the mgmt before the vm_status callback is finished. And we are
returning a *pending* VM status when a `PUT vm_status` task is running.

Added a useful `vm.is_changing_status_to()` that returns a target VM status
when `PUT vm_status` API function is running. The target status is
deducted from the `action` attribute in the apiview dict.
This function can help us to determine whether the VM has already
reached its desired new status.