Hidden input fields for disabled modules due to issue #146 (#173)
Once module is disabled in DC settings we shouldnt display input fields
for module. They werent doing anything so it was just confusing for
user.

* Hidden input fields for disabled modules due to issue #146
* Hiddedn VM forms when snapshots are disabled in issue #146
* Hidden node options for backup and monitoring if disabled issue #146
* Updated changelog
* Closes #146

Node options are checked against global settings as node might be in
multiple DCs, and settings that are set in the main DC are applied.