Renamed Update_ and synchronize_ methods.
The rationale should be that the main method is called synchronize - as we are trying to match the management state with zabbix. Inside that method it is decided, whether the object exists and therefore it is needed just to update it, or it does not and it needs to be created.