Clarified the logic of creation of users and groups
It's a variation of the Chicken or the egg dilemma. After a user group is created, users may be created afterwards with this user group attached. When invoking user creation, new user groups are not permitted to be created.
Also I clarified some ambiguities - group_name clarified to zabbix_group_name as opposed to local_group_name in some places. This will have to be clarified even further so we demonstrate we don't mean host group but user group name.