Implemented Users and Groups one-way synchronization to Zabbix - closes #91 (PR #178)
* Initial placeholders for user/group synchronization hooks, adapter to external alerting configuration #91

* Clarified the logic of creation of users and groups
It's a variation of the Chicken or the egg dilemma. After a user group is created, users may be created afterwards with this user group attached. When invoking user creation, new user groups are not permitted to be created.
Also I clarified some ambiguities - group_name clarified to zabbix_group_name as opposed to local_group_name in some places. This will have to be clarified even further so we demonstrate we don't mean host group but user group name.

* Changed database backend to be able to use transaction hooks.
A separate database backend, coming from django-transaction-hooks is needed before Django 1.9. By using django-transaction-hooks we are able to call celery tasks after db transaction is finished and therefore we eliminate the race condition, when we are trying to get information from the task but it's not in the database yet.