Fixed plotting of longterm graphs
The current implementation produced unsorted data from the Zabbix API,
which was not expected by the JS plotting library and produced strange
graphs in browser. The best solution is to implement sorting in the
`trends.get` API call - https://support.zabbix.com/browse/ZBXNEXT-3974.
In the meantime we will sort the data on the mgmt server.