Updated registration help texts in issue #225
I have simplified texts on the form (two lines of small words) was
barely readable, and added SMS in the text so user can notice.

Updated email texts to have explaining sentence and also moved the
data into "blocks" so it is more visualy more obvious to notice
stars insttead of password so user will notice that password will
be send out by text message (SMS).

Reset password form is automatically submitted by javascript in
order to send password via SMS. We do re-use the Django reset from
that is the reason we have to have the form there and submit.