Disks are defined and therefore we don't create but update disk#247 (#254)
* Disk is already defined and therefore we don't create but update disk. closes #247
* Fixed handling of the disk and nic definition from template. #247
Applying template definition for NICs and VMs should be possible only when VM is defined, not when the definition is updated. Also, when Solaris zones are used, /disk/ representation is very lightweight so it's present automatically with the VM definition creation. Therefore the handling of this special case is different than other /disks/.