Refactored user registration and SMS usage Closes #235
The registration module can now work without SMS.
Text messages will be used for user verification only if the new SMS_REGISTRATION_ENABLED setting is enabled; This should work the same way as it used to be.
With SMS_REGISTRATION_ENABLED=False, new password is requested during registration and password recovery.

This change also modifies the user profile type, address fields to be optional by default. The phone will be optional too if SMS_REGISTRATION_ENABLED=False.

Additionally the SMS framework was refactored so it will be easier to add other SMS providers.