Changed WARNING log level to our custom - IMPORTANT
The WARNING was too confusing.
```
We need this because we want to see some special messages and on node
the default log level of workers is set to WARNING (INFO would be too much noise).
These messages are not errors nor warnings, but can help in some situations (e.g. tracing tasks).
```